:recycle: refactor(candidates): share name/cache helpers via types
stableStringify, clamp, and dropInitials were duplicated as file-private
helpers in candidate-cache and the CA SOS adapter. The upcoming SCC
registrar adapter and scraper need the same stableStringify to build a
byte-identical cache key on both the write and read sides — a mismatch
would be a silent cache miss. Hoist all three into candidate-sources/types
as the single source of truth and import them where they were defined.
